@@ -3,23 +3,41 @@ import 'dart:async';
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:meta/meta.dart';
+import 'package:passline_mobile/crypt/pwgen.dart';
 
 part 'addEdit_event.dart';
 part 'addEdit_state.dart';
 
 class AddEditBloc extends Bloc<AddEditEvent, AddEditState> {
-  AddEditBloc() : super(AddEditState(length: 20));
+  AddEditBloc({name: String, username: String})
+      : super(AddEditState(
+            length: 20,
+            name: name,
+            username: username,
+            password: PwGen.generate(20, true, true, true)));
 
   @override
   Stream<AddEditState> mapEventToState(
     AddEditEvent event,
   ) async* {
-    if (event is AddEditPasswordLength) {
-      yield* _mapPasswordLengthToState(event);
+    if (event is NameChanged) {
+      yield state.copyWith(name: event.name);
+    } else if (event is UsernameChanged) {
+      yield state.copyWith(username: event.username);
+    } else if (event is PasswordLengthChanged) {
+      yield state.copyWith(length: event.length);
+    } else if (event is CharactersSetChanged) {
+      if (state.numbers || state.symbols) {
+        yield state.copyWith(characters: event.characters);
+      }
+    } else if (event is NumbersSetChanged) {
+      if (state.characters || state.symbols) {
+        yield state.copyWith(numbers: event.numbers);
+      }
+    } else if (event is SymbolsSetChanged) {
+      if (state.characters || state.numbers) {
+        yield state.copyWith(symbols: event.symbols);
+      }
     }
-  }
-
-  Stream<AddEditState>_mapPasswordLengthToState(AddEditPasswordLength event) async* {
-    yield AddEditState(length: event.length);
   }
 }
